--- tutorial/kudou/views/html/err.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="jp">
-  <head>
-    <meta charset="utf-8">
-    <title>トップ画面</title>
-  </head>
-  <body>
-    <h1>Error ( *｀ω´)</h1>
-    <a href="./top.html">トップ画面</a>
-  </body>
-</html>
