+++ tutorial/views/html/err.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <h1>Error ( *｀ω´)</h1>
-    <a href="./top.html">トップ画面</a>
+    <button onclick="location.href='./index.html'">目次画面</button>
   </body>
 </html>
