+++ tutorial/views/html/index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="jp">
   <head>
     <meta charset="utf-8">
@@ -8,6 +8,8 @@
   <body>
     <h1>tutorialWebPageIndex</h1>
     <p>tutorialで作成したWEBページのIndexです。</p>
-    <button onclick="location.href='./main.html'">メイン画面</button>
+    <a href="./top.html">top画面</a>
+    <a href="./err.html">エラー画面</a>
+    <a href="./main.html">メイン画面</a>
   </body>
 </html>
