+++ tutorial/views/html/top.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <h1>Hello world ٩( ᐛ )و</h1>
-    <a href="./err.html">エラー画面</a>
+    <button onclick="location.href='./index.html'">目次画面</button>
   </body>
 </html>
